+++ CHANGELOG.md
@@ -6,11 +6,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0]:
+
+* Added support for Polygon:
+
+* Added Covalent explorer for Polygon
+* Added CovalentNFT NFT explorer for Polygon
+
 ## [0.2.0]: 
 
 * Add listing assets support:
   * get a list of Tokens & NFTs associated with a given address
-  * For each assset, populate an object Asset with info on a best effort basis
+  * For each asset, populate an object Asset with info on a best effort basis
 
 * Add Ethplorer explorer for Ethereum
 * Add Blockcypher explorer for Litecoin (& Dogecoin)